Rename Gemfile to default name
When running services in not disabled mode, generated files will have
the name `_service:service-name:file-name`.

We need to use `Gemfile` and `Gemfile.lock` without the prefix as this
is the name bundle used. Since c428978e244b92b5993d7a018ea15de4500671cb
we already find this file, but we need to copy them with the default
name as well.

Co-authored-by: Christian Bruckmayer <cbruckmayer@suse.com>