Work around BUNDLED WITH
With many bundler versions, requiring the class or calling the binary checks
for a file called Gemfile.lock in the current directory, parses that and exits
if the BUNDLED WITH version isn't the bundler version this runs with.

But we have no influence on the bundler version this service is running with,
it can be on anyone's machine. And it also doesn't really matter as we are only
using it to parse/download gems. And BUNDLE WITH is also a deprecated concept.
See https://bundler.io/blog/2015/06/24/version-1-10-released.html

So let's set the BUNDLER_VERSION environment variable which disables all
BUNDLED WITH checking for all bundler versions.